+++ build.js
@@ -65,7 +65,7 @@ const outdir = 'dist/';
 const baseOpts = {
   bundle: true,
   platform: 'node',
-  target: ['node14'],
+  target: ['node16'],
   format: 'cjs',
   external: allDeps,
   watch: opts.dev,